Add HumanEval support via Phase 0.2 conversion and Phase 3.5 integration
Implements HumanEval dataset support through format conversion (Phase 0.2) and seamless integration into existing pipeline (Phase 3.5). Includes attention weights fix, documentation updates, and inspection notebooks.

Phase 0.2: HumanEval to MBPP Conversion
- Convert all 164 HumanEval problems to MBPP format
- Parse test assertions (candidate → function name)
- Output: data/phase0_2_humaneval/humaneval.parquet

Phase 3.5: HumanEval Support
- Update _load_validation_data() to support HumanEval
- Auto-detect dataset and adjust output directories
- Verified: Prompts, code generation, activations, evaluations work correctly

Attention Weights Fix
- Add attn_implementation="eager" to model loading
- Fix regression where SDPA prevented attention extraction
- Required for Phase 6.3 attention pattern analysis

Documentation & Inspection
- Add Pure Manual Config workflow to CLAUDE.md
- Update HUMANEVAL_GEMMA_STEP_BY_STEP.md with decisions and testing
- Defer MBPP backward compatibility testing to LLAMA implementation
- Create file_check_humaneval_conversion.ipynb for Phase 0.2
- Create file_check_humaneval_results.ipynb for Phase 3.5

Technical Changes
- Add dataset_name field to common/config.py
- Add phase0_2_output_dir to config
- Fix float16/bfloat16 conversion in activation saving
- Add Phase 0.2 CLI handler to run.py

+++ common_simplified/activation_hooks.py
@@ -307,7 +307,7 @@ def save_raw_attention_with_boundaries(
     
     np.savez_compressed(
         save_path,
-        raw_attention=attention_tensor.cpu().numpy() if isinstance(attention_tensor, torch.Tensor) else attention_tensor,
+        raw_attention=attention_tensor.cpu().float().numpy() if isinstance(attention_tensor, torch.Tensor) else attention_tensor,
         boundaries=boundaries,
         prompt_length=len(tokenized_prompt.squeeze(0)),
         layer=layer_idx,
